backend/requirements.txt
git+https://gitlab.com/Kononkov1998/rawg-python-wrapper.git@master
git+https://github.com/Kononkov1998/tmdbsimple.git@master
celery==5.4.0
Django>=6.0,<6.1
django-cors-headers>=4.9,<5
django-debug-toolbar>=6.0,<7
djangorestframework>=3.16,<3.17
djangorestframework_simplejwt>=5.5,<5.6
drf-nested-routers>=0.94,<0.95
drf-yasg>=1.21.8,<1.22
Fabric3==1.14.post1
gunicorn==22.0.0
howlongtobeatpy==1.0.17
Pillow==11.0.0
psycopg2==2.9.10
redis==4.5.4
requests==2.32.3
multipledispatch==0.6.0
wheel
python-dotenv
asyncio==3.4.3
uvicorn==0.33.0
adrf>=0.1.9
